Refactor: Switch 이용한 API 통합
